DEV: Remove old migrate_patreon_user_infos OnceOff job
This job was 4 years ago. It's exceptionally unlikely that anyone is still running a version of this plugin which is more than 4 years old, so this code is redundant. It's also incompatible with any future migrations made via the `db/migrate` system. In the extremely unlikely event that this causes an issue for anyone, they can still migrate the data manually on the console